organizing and less endpoint functions to call

+++ spray_backend/views/boulder.py
@@ -52,9 +52,6 @@ def list(request, spraywall_id, user_id):
 @api_view(['POST', 'DELETE'])
 def like_boulder(request, boulder_id, user_id):
     if request.method == 'POST':
-        csrf_token = request.META.get("HTTP_X_CSRFTOKEN")
-        print(csrf_token)
-        print(request.session.get("csrf_token"))
         data = { 'person': user_id, 'boulder': boulder_id }
         like_serializer = LikeSerializer(data=data)
         if like_serializer.is_valid():
@@ -69,7 +66,9 @@ def like_boulder(request, boulder_id, user_id):
     if request.method == 'DELETE':
         # reminder: if liked row is deleted, it automatically deletes the activity row referenced from Like's foreign id in activity. (cascades)
         liked_row = Like.objects.filter(person=user_id, boulder=boulder_id)
-        liked_row.delete()
+        if liked_row.exists():
+            print('DELETING')
+            liked_row.delete()
         data = {
             'isLiked': False
         }
